Refactor form builder and wallet creation components
- Move `showExtra` parameter from form builder view model to view
- Add wallet name collision check in Monero wallet info
- Improve MobX state management in create wallet view
- Fix screen titles not updating (or updating in wrong order)
- Drop FormBuilder dependency on AbstractView (it doesn't play well when nested with mobx)

+++ lib/coins/monero/wallet_info.dart
@@ -77,6 +77,12 @@ class MoneroWalletInfo extends CoinWalletInfo {
     }
     Monero.wPtrList.clear();
     final basePath = p.dirname(walletName);
+    if (File(p.join(basePath, newName)).existsSync()) {
+      throw Exception(Coin.L.error_wallet_name_already_exists);
+    }
+    if (File(p.join(basePath, "$newName.keys")).existsSync()) {
+      throw Exception(Coin.L.error_wallet_name_already_exists);
+    }
     File(walletName).copySync(p.join(basePath, newName));
     File("$walletName.keys").copySync(p.join(basePath, "$newName.keys"));
     // Copy and delete later, if anything throws below we end up with copied walled,
